Updated procedures

+++ public_html/resources/db/create-procedures.sql
@@ -1,5 +1,5 @@
 DELIMITER $$
-CREATE PROCEDURE `AddItemToBasket`(IN user_id INT(11), IN book_id INT(11))
+CREATE PROCEDURE `AddItemToBasket`(IN user_id INT, IN book_id INT)
 BEGIN
 	IF EXISTS (SELECT * FROM basket WHERE basket.user_id = user_id) THEN
 		INSERT INTO basketitem (basket_id, book_id, quantity, cost)
@@ -22,10 +22,20 @@ BEGIN
 		AND books.book_id = book_id;
 	END IF;
 END$$
-DELIMITER ;
 
 DELIMITER $$
-CREATE PROCEDURE `GetBasketByUserId`(IN user_id INT(11))
+CREATE PROCEDURE `EmptyBasket`(IN user_id INT)
+BEGIN
+	DELETE FROM basketitem
+	WHERE basket_id IN
+	(
+		SELECT b.basket_id FROM basket b
+		WHERE b.user_id = user_id
+	);
+END$$
+
+DELIMITER $$
+CREATE PROCEDURE `GetBasketByUserId`(IN user_id INT)
 BEGIN
 	SELECT  bi.book_id,
 			b.title,
@@ -39,10 +49,9 @@ BEGIN
 	WHERE bsk.user_id = user_id
 	GROUP BY book_id;
 END$$
-DELIMITER ;
 
 DELIMITER $$
-CREATE PROCEDURE `GetBookById`(IN bookId INT(11))
+CREATE PROCEDURE `GetBookById`(IN bookId INT)
 BEGIN
 	SELECT  b.book_id,
 			b.title,
@@ -62,7 +71,6 @@ BEGIN
 	WHERE b.book_id = bookId
 	GROUP BY b.book_id;
 END$$
-DELIMITER ;
 
 DELIMITER $$
 CREATE PROCEDURE `GetBooksByCategory`(IN categoryName VARCHAR(50))
@@ -85,23 +93,35 @@ BEGIN
 	WHERE c.name LIKE CONCAT("%", categoryName, "%")
 	GROUP BY b.book_id;
 END$$
-DELIMITER ;
 
 DELIMITER $$
-CREATE PROCEDURE `RemoveItemFromBasket`(IN user_id INT(11), IN book_id INT(11))
+CREATE PROCEDURE `RemoveItemFromBasket`(IN user_id INT, IN book_id INT, IN new_amount INT)
 BEGIN
-	DELETE FROM basketitem
-	WHERE
+	DECLARE remove_limit INT;
+	-- Get the current number of item in the basket
+	SET @current_amount = 
 	(
-		SELECT b.user_id FROM basket b
-		WHERE b.user_id = user_id
+		SELECT count(*) FROM basketitem
+		INNER JOIN basket
+		ON basket.basket_id = basketitem.basket_id
+		WHERE basket.user_id = user_id
 		AND basketitem.book_id = book_id
-	) LIMIT 1;
+	);
+	SET remove_limit = @current_amount - new_amount;
+	IF remove_limit > 0 THEN
+		DELETE FROM basketitem
+		WHERE basket_id IN
+		(
+			SELECT b.basket_id FROM basket b
+			WHERE b.user_id = user_id
+			AND basketitem.book_id = book_id
+		) LIMIT remove_limit;
+	END IF;
 END$$
-DELIMITER ;
 
 GRANT EXECUTE ON PROCEDURE book_store.AddItemToBasket TO 'bs_user'@'localhost';
 GRANT EXECUTE ON PROCEDURE book_store.GetBasketByUserId TO 'bs_user'@'localhost';
 GRANT EXECUTE ON PROCEDURE book_store.GetBookById TO 'bs_user'@'localhost';
 GRANT EXECUTE ON PROCEDURE book_store.GetBooksByCategory TO 'bs_user'@'localhost';
 GRANT EXECUTE ON PROCEDURE book_store.RemoveItemFromBasket TO 'bs_user'@'localhost';
+GRANT EXECUTE ON PROCEDURE book_store.EmptyBasket TO 'bs_user'@'localhost';
